Refactor CatalogItem and related classes to implement Chargeable interface with long return type for unitCharge and unitVat methods

+++ Lab-11/KenaKata/src/main/java/kenakata/catalog/DigitalGood.java
@@ -5,11 +5,11 @@ public class DigitalGood extends StockedGood {
         super();
     }
 
-    public int unitCharge() {
+    public long unitCharge() {
         return 300;
     }
 
-    public int unitVat() {
+    public long unitVat() {
         return 15;
     }
